Added bot ability to respond to tagged messages.
Now, when the bot is tagged in a message in a channel to which it has been added, if link(s) and tagged user(s) are also present in the message, the bot will add the link(s) to the reading list(s) of the tagged user(s). If at least one link or at least one tagged user other than the bot are not present in the message in which the bot is tagged, the bot will reply with an appropriate error message.